requirements.txt
python-telegram-bot==21.6
anthropic==0.34.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-api-python-client==2.143.0
